refactor: general refactor
