+++ gui_components/input_number_control.py
@@ -1,0 +1,18 @@
+from PyQt5.QtWidgets import QLabel, QVBoxLayout, QHBoxLayout, QSpinBox
+
+class InputNumberControl:
+    @staticmethod
+    def setup(label, min, max, value, row, event_callback, spacing=20, control_layout="vertical"):
+        q_label =  QLabel(label)
+        control = QVBoxLayout() if control_layout == 'vertical' else QHBoxLayout()
+        input = QSpinBox()
+        input.setRange(min, max)
+        if value:
+            input.setValue(value)
+        input.valueChanged.connect(event_callback)        
+        control.addWidget(q_label)
+        control.addWidget(input)
+        row.addLayout(control)
+        if spacing:
+            row.addSpacing(20)
+        return control, input
